requirements.txt
Django==1.11
Pygments==2.2.0
djangorestframework==3.6.2
pytz==2017.2
